fixed tests and moved them to test/unit
