+++ handlers/settlements.js
@@ -45,7 +45,7 @@ module.exports = {
         try {
             return await getData
         } catch (e) {
-            console.log(e)
+            throw (Boom.boomify(e))
         }
     }
 };
