+++ handlers/settlements/{settlementId}/participants.js
@@ -42,7 +42,7 @@ module.exports = {
         try {
             return await getData
         } catch (e) {
-            console.log(e)
+            throw (Boom.boomify(e))
         }
     }
 };
