+++ test.xml
@@ -1,0 +1,82 @@
+<?xml version="1.0"?>
+<testsuites>
+  <testsuite tests="0" failures="0" errors="0" name="/settlements"/>
+  <testsuite tests="7" failures="0" errors="0" name="test postSettlementEvent post operation">
+    <testcase name="#1 should be truthy"/>
+    <testcase name="#2 should be truthy"/>
+    <testcase name="#3 Ok response status"/>
+    <testcase name="#4 Ok response status"/>
+    <testcase name="#5 Ok response status"/>
+    <testcase name="#6 Ok response status"/>
+    <testcase name="#7 Ok response status">
+      <system-out>path  settlements.js
+</system-out>
+    </testcase>
+  </testsuite>
+  <testsuite tests="0" failures="0" errors="0" name="/settlements/{settlementId}"/>
+  <testsuite tests="8" failures="0" errors="0" name="test getSettlementsById get operation">
+    <testcase name="#8 should be truthy"/>
+    <testcase name="#9 should be truthy"/>
+    <testcase name="#10 Ok response status"/>
+    <testcase name="#11 Ok response status"/>
+    <testcase name="#12 Ok response status"/>
+    <testcase name="#13 Ok response status"/>
+    <testcase name="#14 Ok response status"/>
+    <testcase name="#15 Ok response status"/>
+  </testsuite>
+  <testsuite tests="0" failures="0" errors="0" name="/settlements/{settlementId}/participants"/>
+  <testsuite tests="8" failures="0" errors="0" name="test getSettledParticipants get operation">
+    <testcase name="#16 should be truthy"/>
+    <testcase name="#17 should be truthy"/>
+    <testcase name="#18 Ok response status"/>
+    <testcase name="#19 Ok response status"/>
+    <testcase name="#20 Ok response status"/>
+    <testcase name="#21 Ok response status"/>
+    <testcase name="#22 Ok response status"/>
+    <testcase name="#23 Ok response status"/>
+  </testsuite>
+  <testsuite tests="0" failures="0" errors="0" name="/settlements/findByParticipant"/>
+  <testsuite tests="8" failures="0" errors="0" name="test getSettlementsByParticipantId get operation">
+    <testcase name="#24 should be truthy"/>
+    <testcase name="#25 should be truthy"/>
+    <testcase name="#26 Ok response status"/>
+    <testcase name="#27 Ok response status"/>
+    <testcase name="#28 Ok response status"/>
+    <testcase name="#29 Ok response status"/>
+    <testcase name="#30 Ok response status"/>
+    <testcase name="#31 Ok response status"/>
+  </testsuite>
+  <testsuite tests="0" failures="0" errors="0" name="/settlementWindows/{settlementWindowId}"/>
+  <testsuite tests="8" failures="0" errors="0" name="test closeSettlementWindow post operation">
+    <testcase name="#32 should be truthy"/>
+    <testcase name="#33 should be truthy"/>
+    <testcase name="#34 Ok response status"/>
+    <testcase name="#35 Ok response status"/>
+    <testcase name="#36 Ok response status"/>
+    <testcase name="#37 Ok response status"/>
+    <testcase name="#38 Ok response status"/>
+    <testcase name="#39 Ok response status"/>
+  </testsuite>
+  <testsuite tests="0" failures="0" errors="0" name="/settlementWindows/findByDateRange"/>
+  <testsuite tests="8" failures="0" errors="0" name="test getSettlementWindowsByDateRange get operation">
+    <testcase name="#40 should be truthy"/>
+    <testcase name="#41 should be truthy"/>
+    <testcase name="#42 Ok response status"/>
+    <testcase name="#43 Ok response status"/>
+    <testcase name="#44 Ok response status"/>
+    <testcase name="#45 Ok response status"/>
+    <testcase name="#46 Ok response status"/>
+    <testcase name="#47 Ok response status"/>
+  </testsuite>
+  <testsuite tests="0" failures="0" errors="0" name="/settlementWindows/findByState"/>
+  <testsuite tests="8" failures="0" errors="0" name="test getSettlementWindowsByState get operation">
+    <testcase name="#48 should be truthy"/>
+    <testcase name="#49 should be truthy"/>
+    <testcase name="#50 Ok response status"/>
+    <testcase name="#51 Ok response status"/>
+    <testcase name="#52 Ok response status"/>
+    <testcase name="#53 Ok response status"/>
+    <testcase name="#54 Ok response status"/>
+    <testcase name="#55 Ok response status"/>
+  </testsuite>
+</testsuites>
