+++ tests/unit/settlements.js
@@ -5,21 +5,21 @@ const Hapi = require('hapi');
 const HapiOpenAPI = require('hapi-openapi');
 const Path = require('path');
 const Mockgen = require('../../data/mockgen.js');
-const responseCodes = [200, 400, 401, 404, 415, 500];
+const responseCodes = [200, 400, 401, 404, 415];
 
 /**
- * Test for /settlementWindows/findByDateRange
+ * Test for /settlements
  */
-Test('/settlementWindows/findByDateRange', function (t) {
+Test('/settlements', function (t) {
 
     /**
-     * summary: Returns Settlement Windows including states and closure reasons. Filtered by date Range.
+     * summary: Trigger the creation of a settlement event, that does the calculation of the net settlement position per participant and marks all transfers in the affected windows as Pending settlement. Returned dataset is the net settlement report for the settlementwindow
      * description:
-     * parameters: startDate, endDate
+     * parameters: settlementEventPayload
      * produces: application/json
      * responses: 200, 400, 401, 404, 415, default
      */
-    t.test('test getSettlementWindowsByDateRange get operation', async function (t) {
+    t.test('test postSettlementEvent post operation', async function (t) {
 
         const server = new Hapi.Server();
         try {
@@ -34,8 +34,10 @@ Test('/settlementWindows/findByDateRange', function (t) {
 
             const requests = new Promise((resolve, reject) => {
                 Mockgen().requests({
-                    path: '/settlementWindows/findByDateRange',
-                    operation: 'get'
+                    path: '/settlements',
+                    operation: 'post',
+                    //response: '200',
+                    //headers: {"Content-Type": "application/json"}
                 }, function (error, mock) {
                     return error ? reject(error) : resolve(mock);
                 });
@@ -48,7 +50,7 @@ Test('/settlementWindows/findByDateRange', function (t) {
             //Get the resolved path from mock request
             //Mock request Path templates({}) are resolved using path parameters
             const options = {
-                method: 'get',
+                method: 'post',
                 url: '/v2' + mock.request.path
             };
             if (mock.request.body) {
@@ -72,10 +74,10 @@ Test('/settlementWindows/findByDateRange', function (t) {
                 t.equal(response.statusCode, responseCode, 'Ok response status');
             }
             t.end();
-
         } catch (e) {
-            console.log(e)
+            t.fail(e)
             t.end()
         }
     });
+
 });
