+++ tests/unit/settlements/findByParticipant.js
@@ -4,37 +4,37 @@ const Test = require('tape');
 const Hapi = require('hapi');
 const HapiOpenAPI = require('hapi-openapi');
 const Path = require('path');
-const Mockgen = require('../../data/mockgen.js');
+const Mockgen = require('../../../data/mockgen.js');
 const responseCodes = [200, 400, 401, 404, 415, 500];
 
 /**
- * Test for /settlementWindows/findByState
+ * Test for /settlements/findByParticipant
  */
-Test('/settlementWindows/findByState', function (t) {
+Test('/settlements/findByParticipant', function (t) {
 
     /**
-     * summary: Returns Settlement Windows including states and closure reasons. Filtered by state.
+     * summary: Returns Settlements per Partricipant (DFSP).
      * description:
-     * parameters: state
+     * parameters: participantId
      * produces: application/json
-     * responses: 200, 400, 401, 404, 415, 500, default
+     * responses: 200, 400, 401, 404, 415, default
      */
-    t.test('test getSettlementWindowsByState get operation', async function (t) {
+    t.test('test getSettlementsByParticipantId get operation', async function (t) {
 
         const server = new Hapi.Server();
         try {
             await server.register({
                 plugin: HapiOpenAPI,
                 options: {
-                    api: Path.resolve(__dirname, '../../config/swagger.json'),
-                    handlers: Path.join(__dirname, '../../handlers'),
+                    api: Path.resolve(__dirname, '../../../config/swagger.json'),
+                    handlers: Path.join(__dirname, '../../../handlers'),
                     outputvalidation: true
                 }
             });
 
             const requests = new Promise((resolve, reject) => {
                 Mockgen().requests({
-                    path: '/settlementWindows/findByState',
+                    path: '/settlements/findByParticipant',
                     operation: 'get'
                 }, function (error, mock) {
                     return error ? reject(error) : resolve(mock);
@@ -65,19 +65,16 @@ Test('/settlementWindows/findByState', function (t) {
             if (mock.request.headers && mock.request.headers.length > 0) {
                 options.headers = mock.request.headers;
             }
-
             for (let responseCode of responseCodes) {
                 server.app.responseCode = responseCode
                 const response = await server.inject(options);
                 t.equal(response.statusCode, responseCode, 'Ok response status');
             }
-
             t.end();
 
         } catch (e) {
-            console.log(e)
+            t.fail(e)
             t.end()
         }
     });
-
 });
