+++ tests/unit/settlements/{settlementId}/participants.js
@@ -4,7 +4,7 @@ const Test = require('tape');
 const Hapi = require('hapi');
 const HapiOpenAPI = require('hapi-openapi');
 const Path = require('path');
-const Mockgen = require('../../../data/mockgen.js');
+const Mockgen = require('../../../../data/mockgen.js');
 const responseCodes = [200, 400, 401, 404, 415, 500];
 
 /**
@@ -26,8 +26,8 @@ Test('/settlements/{settlementId}/participants', function (t) {
         await server.register({
             plugin: HapiOpenAPI,
             options: {
-                api: Path.resolve(__dirname, '../../../config/swagger.json'),
-                handlers: Path.join(__dirname, '../../../handlers'),
+                api: Path.resolve(__dirname, '../../../../config/swagger.json'),
+                handlers: Path.join(__dirname, '../../../../handlers'),
                 outputvalidation: true
             }
         });
@@ -74,7 +74,7 @@ Test('/settlements/{settlementId}/participants', function (t) {
 
         t.end();
     } catch (e) {
-        console.log(e)
+        t.fail(e)
         t.end()
     }
     });
